model/requirements.txt
torch 
torchvision
Pillow
matplotlib
numpy
seaborn 
tqdm 
opencv-python 
scikit-learn
